pyproject.toml
[tool.poetry]
name = "sqlite-to-postgres"
version = "0.1.0"
description = ""
authors = ["lynx <diaamantkia@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
